+++ konlpy/__init__.py
@@ -1,5 +1,5 @@
 __title__ = 'KoNLPy'
-__version__ = '0.3.3-dev'
+__version__ = '0.4.0'
 __author__ = 'Lucy Park'
 __license__ = 'GPL v3'
 __copyright__ = 'Copyright 2014 Lucy Park'