+++ konlpy/jvm.py
@@ -23,12 +23,16 @@ def init_jvm(jvmpath=None):
         return
 
     folder_suffix = ['{0}', '{0}{1}bin',\
-            '{0}{1}jhannanum-0.8.4.jar', '{0}{1}kkma-2.0.jar']
-    javadir = '%s%sjava'\
-            % (utils.installpath, os.sep)
+            '{0}{1}jhannanum-0.8.4.jar',\
+            '{0}{1}kkma-2.0.jar',\
+            '{0}{1}komoran-2.4-e.jar',\
+            '{0}{1}shineware-common-2.0.jar', '{0}{1}shineware-ds-1.0.jar',\
+            '{0}{1}*']
+
+    javadir = '%s%sjava' % (utils.installpath, os.sep)
+
     args = [javadir, os.sep]
     classpath = os.pathsep.join(f.format(*args) for f in folder_suffix)
 
     jvmpath = jvmpath or jpype.getDefaultJVMPath()
-    jpype.startJVM(jvmpath, '-Djava.class.path=%s' % classpath, '-ea',\
-                            '-Xmx768m')
+    jpype.startJVM(jvmpath, '-Djava.class.path=%s' % classpath, '-ea', '-Xmx768m')