+++ konlpy/tag/__init__.py
@@ -1,5 +1,6 @@
 from ._hannanum import Hannanum
 from ._kkma import Kkma
+from ._komoran import Komoran
 try:
     from ._mecab import Mecab
 except ImportError: